fix?: dev: Github thinks that the article was written in the future
The actions pipeline skips generating the HTML for this post cause it
thinks it was written in the future:

```
Skipping: _posts/2022-06-19-halmos-writing.liquid.md has a future date
```

This is maybe cause the servers running the actions are in the USA?
and I'm in Germany, a few hours ahead.

I'm setting the date to yesterday, hopefully that will work.